% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_object.R
\name{chlaa_as_fit}
\alias{chlaa_as_fit}
\title{Coerce an object into class `chlaa_fit`}
\usage{
chlaa_as_fit(x)
}
\arguments{
\item{x}{A fit object (typically from `chlaa_fit_pmcmc()`).}
}
\value{
The same object with class `chlaa_fit` prepended.
}
\description{
Coerce an object into class `chlaa_fit`
}
