% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{chlaa_fit_select_iterations}
\alias{chlaa_fit_select_iterations}
\title{Select iterations from a posterior draws matrix}
\usage{
chlaa_fit_select_iterations(draws, burnin = 0.5, thin = 1)
}
\arguments{
\item{draws}{Matrix of posterior draws (iterations x parameters).}

\item{burnin}{Burn-in, either proportion in (0,1) or an integer count.}

\item{thin}{Thinning interval (integer >= 1).}
}
\value{
A matrix subset of draws.
}
\description{
Select iterations from a posterior draws matrix
}
