% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{chlaa_parameter_info}
\alias{chlaa_parameter_info}
\title{Parameter metadata table for the cholera model}
\usage{
chlaa_parameter_info()
}
\value{
A data.frame with columns: name, default, units, description, source.
}
\description{
Returns metadata for all parameters required by the odin model and observation model.
Defaults here are aligned to the defaults in `inst/odin/cholera_model.R`
}
