% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{chlaa_parameters_validate}
\alias{chlaa_parameters_validate}
\title{Validate a cholera parameter list}
\usage{
chlaa_parameters_validate(pars)
}
\arguments{
\item{pars}{Parameter list.}
}
\value{
Invisibly returns `pars` if valid, otherwise errors.
}
\description{
Validate a cholera parameter list
}
