% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{chlaa_plot_forecast}
\alias{chlaa_plot_forecast}
\title{Plot a forecast summary}
\usage{
chlaa_plot_forecast(
  forecast,
  var,
  data = NULL,
  data_time = "time",
  data_y = NULL,
  show_mean = FALSE
)
}
\arguments{
\item{forecast}{Output from `chlaa_forecast_from_fit()`.}

\item{var}{Variable to plot.}

\item{data}{Optional observed data frame (must have a time column and a matching y column).}

\item{data_time}{Column name for time in `data`.}

\item{data_y}{Column name for y in `data`. If NULL, uses `var`.}

\item{show_mean}{Plot mean line as well as median.}
}
\value{
A ggplot object.
}
\description{
Plot a forecast summary
}
