% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{chlaa_plot_incidence}
\alias{chlaa_plot_incidence}
\title{Plot daily incidence summary for a variable}
\usage{
chlaa_plot_incidence(sim, var = "inc_symptoms")
}
\arguments{
\item{sim}{Simulation output.}

\item{var}{Incidence variable (default inc_symptoms).}
}
\value{
A ggplot object.
}
\description{
Plot daily incidence summary for a variable
}
