% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_diagnostics_extra.R
\name{chlaa_plot_likelihood_density}
\alias{chlaa_plot_likelihood_density}
\title{Plot Distribution Of Likelihood Values}
\usage{
chlaa_plot_likelihood_density(fit, burnin = 0, thin = 1)
}
\arguments{
\item{fit}{Fit object returned by `chlaa_fit_pmcmc()`.}

\item{burnin}{Burn-in proportion in (0,1) or iteration count.}

\item{thin}{Thinning interval.}
}
\value{
A ggplot object.
}
\description{
Plot Distribution Of Likelihood Values
}
