% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{chlaa_scenario}
\alias{chlaa_scenario}
\title{Create a counterfactual scenario}
\usage{
chlaa_scenario(name, modify)
}
\arguments{
\item{name}{Scenario name.}

\item{modify}{Named list of parameter modifications.}
}
\value{
An object of class `chlaa_scenario`.
}
\description{
Create a counterfactual scenario
}
