% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual_grid.R
\name{chlaa_scenarios_from_grid}
\alias{chlaa_scenarios_from_grid}
\title{Convert a counterfactual grid to scenario objects}
\usage{
chlaa_scenarios_from_grid(grid)
}
\arguments{
\item{grid}{Output from `chlaa_counterfactual_grid()`.}
}
\value{
A list of `chlaa_scenario` objects.
}
\description{
Convert a counterfactual grid to scenario objects
}
