% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{chlaa_simulate_from_snapshot}
\alias{chlaa_simulate_from_snapshot}
\title{Simulate Forward From A Snapshot}
\usage{
chlaa_simulate_from_snapshot(
  snapshot,
  pars = NULL,
  time,
  n_threads = 1,
  deterministic = FALSE
)
}
\arguments{
\item{snapshot}{Snapshot object from `chlaa_snapshot_create()`.}

\item{pars}{Optional parameter list to use from snapshot onward.}

\item{time}{Times to simulate to (must be >= `snapshot$time`).}

\item{n_threads}{Threads for dust2.}

\item{deterministic}{Deterministic mode (if supported).}
}
\value{
Simulation output data.frame as in `chlaa_simulate()`.
}
\description{
Simulate Forward From A Snapshot
}
