% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_summary.R
\name{chlaa_standard_scenarios}
\alias{chlaa_standard_scenarios}
\title{Build standard operational scenario templates}
\usage{
chlaa_standard_scenarios(
  pars,
  trigger_time,
  horizon,
  vax_total_doses = 280000,
  early_offset = -14,
  late_offset = 14,
  campaign_days = 6,
  wash_duration = 120,
  care_duration = 120
)
}
\arguments{
\item{pars}{Baseline parameters.}

\item{trigger_time}{Trigger time for anticipatory interventions.}

\item{horizon}{End of horizon (used to cap intervention windows).}

\item{vax_total_doses}{Total vaccine doses for vaccination scenarios.}

\item{early_offset}{Days relative to trigger for early vaccination start.}

\item{late_offset}{Days relative to trigger for late vaccination start.}

\item{campaign_days}{Vaccination campaign duration.}

\item{wash_duration}{WASH intervention duration.}

\item{care_duration}{Care intervention duration.}
}
\value{
A list of `chlaa_scenario` objects.
}
\description{
Build standard operational scenario templates
}
