% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_object.R
\name{cholera_as_fit}
\alias{cholera_as_fit}
\title{Coerce an object into class `cholera_fit`}
\usage{
cholera_as_fit(x)
}
\arguments{
\item{x}{A fit object (typically from `cholera_fit_pmcmc()`).}
}
\value{
The same object with class `cholera_fit` prepended.
}
\description{
Coerce an object into class `cholera_fit`
}
