% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_object.R
\name{cholera_fit_metadata}
\alias{cholera_fit_metadata}
\title{Extract fit metadata (packer, prior, start parameters, data)}
\usage{
cholera_fit_metadata(fit)
}
\arguments{
\item{fit}{A `cholera_fit` object.}
}
\value{
A named list.
}
\description{
Extract fit metadata (packer, prior, start parameters, data)
}
