% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{cholera_parameters_print}
\alias{cholera_parameters_print}
\title{Print a parameter table}
\usage{
cholera_parameters_print(pars)
}
\arguments{
\item{pars}{Parameter list.}
}
\value{
A data.frame joining values to metadata.
}
\description{
Print a parameter table
}
