% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cholera_plot_ce_plane}
\alias{cholera_plot_ce_plane}
\title{Plot a cost-effectiveness plane from scenario comparisons}
\usage{
cholera_plot_ce_plane(cmp)
}
\arguments{
\item{cmp}{Comparison table.}
}
\value{
A ggplot object.
}
\description{
Expects output from `cholera_compare_scenarios(..., include_econ = TRUE)`.
}
