% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_workflow.R
\name{cholera_plot_trace}
\alias{cholera_plot_trace}
\title{Plot fit traces for selected parameters}
\usage{
cholera_plot_trace(fit, parameters = NULL, burnin = 0, thin = 1)
}
\arguments{
\item{fit}{A `cholera_fit` object.}

\item{parameters}{Optional subset of parameter names.}

\item{burnin}{Burn-in proportion or count.}

\item{thin}{Thinning interval.}
}
\value{
A ggplot object.
}
\description{
Plot fit traces for selected parameters
}
