% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cholera_plot_trajectories}
\alias{cholera_plot_trajectories}
\title{Plot trajectories of selected model variables}
\usage{
cholera_plot_trajectories(
  sim,
  vars = c("inc_symptoms", "cum_deaths"),
  facet = TRUE
)
}
\arguments{
\item{sim}{Output from `cholera_simulate()` or `cholera_run_scenarios()`.}

\item{vars}{Variables to plot.}

\item{facet}{If TRUE, facet by variable.}
}
\value{
A ggplot object.
}
\description{
Plot trajectories of selected model variables
}
