% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{cholera_scenario}
\alias{cholera_scenario}
\title{Create a counterfactual scenario}
\usage{
cholera_scenario(name, modify)
}
\arguments{
\item{name}{Scenario name.}

\item{modify}{Named list of parameter modifications.}
}
\value{
An object of class `cholera_scenario`.
}
\description{
Create a counterfactual scenario
}
