% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual_grid.R
\name{cholera_scenarios_from_grid}
\alias{cholera_scenarios_from_grid}
\title{Convert a counterfactual grid to scenario objects}
\usage{
cholera_scenarios_from_grid(grid)
}
\arguments{
\item{grid}{Output from `cholera_counterfactual_grid()`.}
}
\value{
A list of `cholera_scenario` objects.
}
\description{
Convert a counterfactual grid to scenario objects
}
